server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai-whisper==20231117
groq==0.4.1
google-generativeai==0.3.0
Pillow==10.0.0
python-dotenv==1.0.0
numpy==1.24.3
transformers==4.35.0
torch==2.0.1
pdf2image==1.16.3
pytesseract==0.3.10
pyttsx3==2.90
gTTS==2.4.0
aiofiles==23.2.1
